Fix: signup required form 수정
수정과 더불어 파일 위치가 바뀌었습니다.
