+++ src/components/signup/signupForm/styles.ts
@@ -1,0 +1,19 @@
+import styled from '@emotion/styled';
+//import { colors } from '@/styles/colorPalette';
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 0 170px;
+`;
+
+export const Blank = styled.div`
+  margin: 25px;
+`;
+
+export const Form = styled.form`
+  width: 600px;
+  margin: 80px 0;
+`;
