--- src/components/signupForm/styles.ts
@@ -1,18 +1,0 @@
-import styled from '@emotion/styled';
-//import { colors } from '@/styles/colorPalette';
-
-// 임시 --> 추후 삭제
-export const ShadowBox = styled.div`
-  width: 940px;
-  height: 1000px;
-  border-radius: 30px;
-  background: #fff;
-  box-shadow: 0px 0px 30px 0px rgba(106, 106, 106, 0.15);
-  z-index: 1;
-  display: flex;
-  justify-content: center;
-`;
-
-export const Blank = styled.div`
-  margin: 25px;
-`;
